Shooter tweaks
halved speed tolerance for the shooter.
Setup a seperate time control for the "power control" period during
shots
Added a slow firing power control routine for testing.
Tweaked autonomous RPMs for 7shot (seperate RPM for first 3 and last
four)
Added some code to allow the last shot in a group set a new RPM, not
used.
Seperated the "fire" commands at the end of 7 shot, and added a delay
between the last 4